pyproject.toml
[project]
name = "byte-pair-encoding-implementations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.28.0",
    "pyarrow>=18.0.0",
    "pyyaml>=6.0",
]
